Chore: transactional 제거
