+++ Admin/src/main/java/com/seveneleven/company/service/CompanyFileHistoryService.java
@@ -4,9 +4,8 @@ import com.seveneleven.entity.file.FileMetadata;
 import org.springframework.transaction.annotation.Transactional;
 
 public interface CompanyFileHistoryService {
-    @Transactional
+
     void registerLogoImageHistory(FileMetadata uploadedFileEntity, Long registrantId);
 
-    @Transactional
     void deleteLogoImageHistory(FileMetadata deletedFileEntity, Long registrantId);
 }
