+++ Admin/src/main/java/com/seveneleven/company/service/CompanyFileService.java
@@ -5,12 +5,10 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
 public interface CompanyFileService {
-    @Transactional
+
     void uploadLogoImage(MultipartFile file, Long companyId, Long uploaderId);
 
-    @Transactional(readOnly = true)
     FileMetadataResponse getLogoImage(Long companyId);
 
-    @Transactional
     void deleteLogoImage(Long companyId, Long deleterId);
 }
