+++ Main/src/main/java/com/seveneleven/board/service/PostFileHistoryService.java
@@ -4,9 +4,8 @@ import com.seveneleven.entity.file.FileMetadata;
 import org.springframework.transaction.annotation.Transactional;
 
 public interface PostFileHistoryService {
-    @Transactional
+
     void registerPostFileHistory(FileMetadata uploadedFileEntity, Long registrantId);
 
-    @Transactional
     void deletePostFileHistory(FileMetadata deletedFileEntity, Long registrantId);
 }
