+++ Main/src/main/java/com/seveneleven/board/service/PostFileService.java
@@ -7,15 +7,12 @@ import org.springframework.web.multipart.MultipartFile;
 import java.util.List;
 
 public interface PostFileService {
-    @Transactional
+
     void uploadPostFiles(List<MultipartFile> files, Long postId, Long uploaderId);
 
-    @Transactional(readOnly = true)
     List<FileMetadataResponse> getPostFiles(Long postId);
 
-    @Transactional
     void deletePostFile(Long postId, Long fileId, Long deleterId);
 
-    @Transactional
     void deleteAllPostFiles(Long postId, Long deleterId);
 }
