+++ Main/src/main/java/com/seveneleven/board/service/PostLinkService.java
@@ -7,15 +7,12 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 
 public interface PostLinkService {
-    @Transactional
+
     void uploadPostLinks(List<LinkInput> linkInputs, Long postId, Long uploaderId);
 
-    @Transactional(readOnly = true)
     List<LinkResponse> getPostLinks(Long postId);
 
-    @Transactional
     void deletePostLink(Long postId, Long linkId, Long deleterId);
 
-    @Transactional
     void deleteAllPostLinks(Long postId);
 }
